requirements.txt
Flask==2.2.5
opencv-python==4.12.0.88
numpy==2.1.3
Pillow==11.0.0
Werkzeug==2.2.3
requests==2.32.3
gunicorn==20.1.0